+++ CommonServer/Services/TeamMemberService.ts
@@ -15,6 +15,10 @@ import DeleteBy from '../Types/Database/DeleteBy';
 import ObjectID from 'Common/Types/ObjectID';
 import QueryHelper from '../Types/Database/QueryHelper';
 import LIMIT_MAX from 'Common/Types/Database/LimitMax';
+import ProjectService from './ProjectService';
+import { IsBillingEnabled } from '../Config';
+import BillingService from './BillingService';
+import SubscriptionPlan from 'Common/Types/Billing/SubscriptionPlan';
 
 export class Service extends DatabaseService<Model> {
     public constructor(postgresDatabase?: PostgresDatabase) {
@@ -70,6 +74,10 @@ export class Service extends DatabaseService<Model> {
             onCreate.createBy.data.userId!,
             onCreate.createBy.data.projectId!
         );
+
+
+        await this.updateSubscriptionSeatsByUnqiqueTeamMembersInProject(onCreate.createBy.data.projectId!);
+        
         return createdItem;
     }
 
@@ -129,9 +137,55 @@ export class Service extends DatabaseService<Model> {
     ): Promise<OnDelete<Model>> {
         for (const item of onDelete.carryForward as Array<Model>) {
             await this.refreshTokens(item.userId!, item.projectId!);
+            await this.updateSubscriptionSeatsByUnqiqueTeamMembersInProject(item.projectId!);
         }
 
         return onDelete;
+    }
+
+    public async getUniqueTeamMemberCountInProject(projectId: ObjectID): Promise<number> { 
+        return (await this.countBy({
+            query: {
+                projectId
+            },
+            props: {
+                isRoot: true
+            },
+            distinctOn: "userId",
+            skip: 0,
+            limit: LIMIT_MAX
+        })).toNumber()
+    }
+
+    public async updateSubscriptionSeatsByUnqiqueTeamMembersInProject(projectId: ObjectID): Promise<void> {
+        
+        if (!IsBillingEnabled) {
+            return;
+        }
+        
+        const numberOfMembers = await this.getUniqueTeamMemberCountInProject(projectId);
+        const project = await ProjectService.findOneById({
+            id: projectId,
+            select: {
+                paymentProviderSubscriptionId: true,
+                paymentProviderPlanId: true
+            },
+            props: {
+                isRoot: true
+            }
+        });
+
+        
+
+        if (project && project.paymentProviderSubscriptionId && project?.paymentProviderPlanId) {
+            const plan = SubscriptionPlan.getSubscriptionPlanById(project?.paymentProviderPlanId!);
+            
+            if (!plan) {
+                return;
+            }
+
+            await BillingService.updateSubscription(project.paymentProviderSubscriptionId, plan, numberOfMembers, plan.getYearlyPlanId() === project.paymentProviderPlanId);
+        }
     }
 }
 export default new Service();
